style(BACK): Reorganize modules
