+++ back/src/modules/cdn/cdn.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from '../users/users.service';
+import { UsersModule } from '../users/users.module';
 import { CdnController } from './cdn.controller';
 
 @Module({
-	imports: [],
-	controllers: [CdnController],
-	providers: [UsersService],
+    imports: [UsersModule],
+    controllers: [CdnController],
+    providers: [],
 })
 export class CdnModule {}
