+++ back/src/modules/chat/chat.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ChatGateway } from '../chat-gateway/chat.gateway';
+import { ChatGateway } from './chat.gateway';
 import { ChatService } from './chat.service';
 
 @Module({ providers: [ChatGateway, ChatService] })
