+++ back/src/modules/prisma/prisma.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PrismaService } from './prisma.service';
 
 @Module({
-	providers: [PrismaService],
-	exports: [PrismaService], // 👈 export PrismaService for DI
+    providers: [PrismaService],
+    exports: [PrismaService],
 })
 export class PrismaModule {}
