+++ back/src/modules/users/users.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
+import { PrismaModule } from '../prisma/prisma.module';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 
 @Module({
-	imports: [],
-	controllers: [UsersController],
-	providers: [UsersService],
+    imports: [PrismaModule],
+    controllers: [UsersController],
+    providers: [UsersService],
+    exports: [UsersService],
 })
 export class UsersModule {}
